Se hace refactorizacion para mejorar la legibilidad y la mantenibilidad del código.
