+++ src/components/translator/navbar/NavBar.css
