+++ src/components/translator/navbar/NavBar.tsx
