+++ src/components/translator/sections/styles/Script.css
