+++ src/components/translator/sections/styles/Video.css
@@ -6,15 +6,15 @@
 }
 
 .video-player-container video {
-  display: block;
+  display: flex;
   height: auto;
   max-width: 100%;
 }
 
 .video-controls {
   display: flex;
-  gap: clamp(5px, 1vw, 15px);
-  margin-top: clamp(5px, 0.8vw, 10px);
+  gap: clamp(2px, 0.5vw, 6px);
+  margin-top: clamp(2px, 0.4vw, 5px);
 }
 
 .video-controls button:hover {
@@ -34,9 +34,9 @@
   color: white;
   cursor: pointer;
   display: flex;
-  font-size: clamp(12px, 1.2vw, 16px);
+  font-size: clamp(5px, 0.5vw, 7px);
   justify-content: center;
-  padding: clamp(4px, 0.8vw, 10px) clamp(8px, 1.5vw, 15px);
+  padding: clamp(1px, 0.3vw, 4px) clamp(3px, 0.6vw, 6px);
   transition: background-color 0.2s;
 }
 
@@ -51,7 +51,7 @@
 }
 
 .video-container {
-  display: flex;
+  display: grid;
   flex-direction: column;
   outline: none;
   width: clamp(300px, 100%, 1200px);
@@ -63,13 +63,16 @@
 
 .video-progress {
   margin: clamp(10px, 1.2vw, 20px) 0;
+  overflow: hidden;
+  max-width: 100%;
+  box-sizing: border-box;
 }
 
 .keyboard-controls-help {
   align-items: center;
   color: #666;
   display: flex;
-  font-size: clamp(12px, 1.2vw, 16px);
+  font-size: clamp(19px, 1.2vw, 16px);
   margin-left: auto;
 }
