+++ src/components/translator/sections/styles/styles.ts
