+++ src/interface/interfaces.ts
