+++ src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import { EditionPage } from "./components/principalpages/EditionPage.tsx";
+import { EditionPage } from "./EditionPage";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
